refactor: Decompose PWA installer into hook
