+++ src/frontend/.gitignore
@@ -1,4 +1,4 @@
 node_modules
 .next
 next-env.d.ts
-public/*.js*
+public/*.js
